Moved hasServiceNameInCaseTypeId to CallbackValidator
